function that prints all the elements of a listint_t list.

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -2,12 +2,12 @@
 #include "lists.h"
 /**
  * print_listint - prints all the elements of a list
- * @h: list
+ * @h: pointer to the head of the list
  * Return: the number of nodes
  */
 size_t print_listint(const listint_t *h)
 {
-	size_t nodes = 0;
+	int nodes = 0;
 
 	while (h != NULL)
 	{
